Fixed admin notices after changing alert settings
Up until now the plugin was displaying an alert message saying that it
was able to write the new alert settings either in the database or the
plain text file (introduced with version 1.7.18) but if the database or
the file is not available the plugin is failing silently and still
reporting the changes applied in the alert settings page. Now the plugin
checks if the setting was actually saved before it considers the
submission a success.