Fixed rare edge cases with the hardening of core directories
When the user decides to revert the hardening in one of the core folder
the plugin was supposed to delete the access control file if it detected
that the file was empty after the deletion of the denying rules, this
was not happening, this commit adds the missing feature. Additionally,
the hardening of a directory with an existing access control file was
breaking the website with a 500 Internal Server Error due to missing new
line characters in the existing rules, we fixed this by forcing the
plugin to append a new line character right before any pre-existing
content of the file.