Fix format of the README file with new WordPress design
WordPress plugins market place changed its design and now the line
ending from the readme files are translated to HTML break tags, this
causes some of the text in the current paragraphs to over-flow. This
commit fixes this by merging the lines of each paragraph into one as
normal Markdown file.