Modify structure of the project for maintainability
Up until now the development of the plugin has been managed using a set
of PHP files stored in a non-tracked directory. These files were being
merged all together into a giant PHP file via some Makefile instructions
that are not available anymore. This is, to allow future contributors to
collaborate in the development of the code without fear of regression
during the modification of a file that was +12k LoC.