Remove minified files to facilitate future contributions
Up until now the source code for the plugin was being generated with a
set of instructions defined in a Makefile, this included some actions to
minify the CSS and JavaScript files. However, external contributions did
not have access to the Makefile so we have decided to revert the
minification to facilitate the open-source contributions.