Add restriction to prevent direct access to PHP files
After we split the giant PHP files from previous versions of the plugin
into small files in a subdirectory we became concerned with the security
of this approach as malicious users could attempt to access these files
directly to cause a FPD [1]. This commit fixes this possible issue by
sending a 403/Forbidden HTTP status code if the code is not executed in
conjunction with the initializer which is only loaded by WordPress from
the admin dashboard.

Notice that this was already fixed with some rules in the Apache access
control file distributed with every release of the code, but for people
using a different web server it makes sense to add this additional layer
of protection.

[1] https://www.owasp.org/index.php/Full_Path_Disclosure