Fix ajax request handlers with new project structure
The Ajax requests are being handled by multiple functions associated to
each main page. The function names are attached to the WordPress action
triggers from the PHP file with the global variables, but this file is
being loaded before the functions are defined. This commit fixes this
issue by loading the file with the global variables and action triggers
at the end of everything.