Add HTTP request parameters to track some settings
As part of our plan to deprecate the Socket interface introduced a while
ago to support websites without the CURL extension, we are going to
start sending the name of the HTTP transport mechanism. After some time,
if we detect that the number of websites using Sockets is too low to
keep supporting that version of the code we will remove it. Also, we
will report the status of the SSL certificate verification, we want to
force every website to communicate via HTTPS.