Fix length of the pagination helper with many pages
The commit shortens the pagination buttons when the maximum number of
pages is smaller than the number of items in the page. With this change,
a pagination with 100 buttons will be reduced to 20 and when the user
requests a page with a number higher than the middle the starting point
will be moved to the left, so if the user goes to page 12 the starting
point will be 3 instead of 1 and the final button will be moved forward
too, in this case to 22 instead of 20.