Modify entire code base to enforce HTTPS over HTTP
Up until now the plugin has been providing some options to disable the
SSL certificate verification in servers where the OpenSSL library is too
old or the CURL extension was compiled without support for algorithms
used in the SSL certificates installed in the API service. We will start
enforcing HTTPS and display a warning to website owners when the SSL
verification is not available.