Remove heartbeat settings after performance improvement
The heartbeat section in the settings page was intended to help website
owners to configure the intervals in which WordPress sends Ajax requests
to control access to certain features when multiple accounts are doing
the same things like editing the same page, post, file, comment, etc.
This was necessary in older versions of the plugin because the XHR
monitor was affecting the responsiveness of many websites since it runs
a heavy filter against every Ajax request. Newer versions of the code
have improved this filter so the heartbeat settings are now useless.