Modify failed logins logger with wrong passwords
This will force the plugin to log the passwords during a failed login
attempt. This is useful for the website owner if he suspects of a brute
force attack and will let them know what technique is being used to
attack the login page. Up until now the password logger was behind an
option that was disabled by default, this was because if a legitimate
user mistypes his password the plugin will still send it to the API
which may become a security hole if a malicious user knows the API key
as they can retrieve the audit logs and see the password there. After
some thinking we decided to enable this feature by default and just log
the password in the local database instead of the remote API service.