Add tool to import and export the plugin settings
To prevent unwanted arbitrary writes into disk, the import and export
feature will only process allowed option names, these options are
selected manually by one of the developers. Any other option, valid or
not will be ignored. Notice that the export function also ignores any
option that is present in the JSON file by not whitelisted by the
developer in the code.