Fix plugin reinstall procedure with backup and prechecks
This commit modifies the tool that allows website owners to reset one or
more plugins at once. The tool will now create a backup of the existing
plugin before it proceeds with the installation of a fresh copy gotten
from the official WordPress repositories. If the process fails in any of
the steps the code will revert the backup to its original location.