Fix storage folder creation with WPCLI for issue #25
Up until now the creation of the storage folder has been triggered by
the access of any of the admin pages. However, for some power users who
need to manage multiple websites using tools like WPCLI this process is
unavailable as they never access the admin dashboard. This commit fixes
this by allowing them to trigger the creation of the storage folder by
sending a HTTP request to the login page with invalid credentials, this
will trigger the log for the failed login event and subsequently the
creation of the storage folder.