Add message in the core integrity tool for false/positives
Files marked as fixed in the core integrity checks are moved to a cache
file that is used to skip some of the scans. This becomes a vulnerability
if these files are later infected with malicious code. This commit adds
a warning message advising the website owner to review the files marked
as false/positives to be sure that the website is completely clean.