Add mechanism to display self-hosting logs as fallback
If the website doesn't has a valid API key but the self-hosting monitor
is enabled, the plugin should read the logs from that file. To reduce
the memory consumption it will only read the latest lines inserted in
the file, this also allows us to control the limits so the plugin keeps
the file reader boudaries in place.