Add option to display differences in core integrity checks
For advanced users the plugin will allow them to view the differences
found by the plugin during the execution of the WordPress integrity
checks. This is done using the Unix diff utility, by comparing the
content of the original code distributed with every installation package
and the content of the file that is currently in the server. Notice that
if the hosting provider has blocked the execution of external commands
this tool will not work because it relies on the availability of the PHP
exec method to execute diff against the two files.