Fix static data storage path to allow server migrations
Unless the website owners sets the constant SUCURI_DATA_STORAGE into the
configuration file, the plugin will use a semi-hardcoded directory path
to store its settings, security logs, cache and additional data. This
was causing some issues with certain hosting providers because after a
migration the directory path (which was being set as an additional
option in the settings file) was not being changed dynamically. This
commit fixes this by removing the static option from the settings file
and instead generating the directory path dynamically.