Add option to ignore non-registered custom post-types
If the website owner is receiving alerts for post types that are not
listed in the "Ignore Post Changes" setting it may be because the theme
or plugin that is making these changes is registering the custom
post-type on runtime, in this case our plugin will not be able to detect
these changes and consequently you will not be able to ignore those
alerts. However, if the website owner knows the unique identifier of the
post-type they can type it in this new form and our plugin will do its
best to skip the alerts associated to that.