Add more details into the event that monitors post deletions
Every time a post or page is deleted WordPress triggers an action called
delete_post, this action is caught by the event monitor. However, at
this point the page or post has already been deleted, which means we
cannot send enough information about the event to the API. To fix this,
we will also monitor the before_delete_post action which is triggered
before the post is deleted but after the user has executed the action.
We will store some information related to the post in a temporary data
structure. Then, when the delete_post action is triggered we will
extract this informaiton to send it to the API. We will delete the
temporary data after the operation has succeeded.