Add one-time newsletter invitation after plugin updates
For every fresh installation of the plugin we will send a one-time email
to the website owner with an invitation to subscribe to our security
related newsletter where they can learn about better security practices
and get alerts from public vulnerabilities disclosures. The plugin will
set a hidden (non-modifiable) option in the settings file to let it know
that we already sent this email. In most cases this will be enough to
avoid unnecessary spam, but if the website owner decides to reset the
plugin settings, or for some reason, the settings file is not writable.