Remove option to configure the file scanner algorithm
Up until now the plugin allowed the website owner to choose which
algorithm would be used to scan the directory tree, the options included
the SPL library, OpenDIR and Glob. With the pass of the years and the
release of PHP 7 many hosting providers upgraded their systems to newer
versionf of PHP to get advantage of the performance boost of the latest
features. This means that most of the biggest hosting providers in the
market are using at least PHP5 which includes SPL so there is no need to
choose between the other algorithms with less performance. A warning
will appear in the plugin dashboard if the SPL library is not available.