Fix the invitation to subscribe to our newsletter
When the plugin detects that the version number of the plugin that was
previously installed is different than the version that is currently
installed, it assumes that the plugin has been updated. When this
happens it displays a message in the admin dashboard inviting the users
to subscribe to our newsletter. However, in some rare cases this fails
because the HTTP headers has already been sent and the plugin throws an
error. This commit fixes this problem by checking if the HTTP headers
has already been sent before the alert is printed.