Fix language file loader when locale is not supported
The plugin will check which language the website is to configured to
use, if the plugin has no support for this language it will try to copy
the English POT files (both the raw file and the compiled version) to a
file with the same locale as the website. If the files cannot be copied
the plugin will force the global locale variable to use English.