Add option to ignore events for post transitions
When the type of a post is changed, for example, from new to publish or
from draft to private, or from private to trash, etc the plugin will
send an email alert to the admins of the website and also report the
change to the API. However, in some cases these transitions happen too
often, specially with the draft status, and people may prefer to ignore
such alerts. This commit adds an option to the settings page to allow
people to disable the email alerts for post status transitions.