Fix infinite loop with email alerts and SMTP plugin
The plugin detects changes in the posts, there are some other plugins
that intercept PHPMailer and create a post object that is later used to
send the real message to the users. This object is also detected by our
plugin and is considered an additional event that must be reported, so
after the first execution the operation falls into an infinite loop.