Add mechanism to automatically blacklist IP addresses
For every failed login attempt the plugin will log the event into a flat
file, once five failed logins have beem recorded the plugin will send
the IP address to the Firewall IP to be blacklisted. If the operation
succeeds the plugin will delete the records from the flat file to free
space. If the operation fails, it will keep counting the failed logins
into the flat file until the Firewall API is responsive. Both the
blacklist and unblacklist actions will be logged into the event monitor
as well for visibility.