Add option to stop sending the failed login passwords
By default, the plugin will report which password was used during a
password guessing brute force attack, both to the API service and to the
admin users via an email alert. This may pose a security breach if one
of the admin users missed a character in their password and mistakingly
sent the text to the API and to the other administrators. For this, we
will allow the admins to disable this behavior.