Add option to configure the malware scanner target URL
The remote malware scanner provided by the plugin is powered by Sucuri
SiteCheck, a service that takes a publicly accessible URL and scans it
for malicious code. If your website is not visible to the Internet, for
example, if it is hosted in a local development environment or a
restricted network, the scanner will not be able to work on it. Also,
if the website was installed in a non-standard directory the scanner
will report a "404 Not Found" error. You can use this option to change
the URL that will be scanned.