Add option to enable the auto clear cache firewall function
Every time a page or post is modified and saved into the database the
plugin will send a HTTP request to the firewall API service and except
that, if the API key is valid, the cache is reset. Notice that the cache
of certain files is going to stay as it is due to the configuration on
the edge of the servers.