Add smart limit to send logs from the queue to the API
We will use the maximum execution time setting to limit the number of
logs that the plugin will try to send to the API service before the
server times out. In a regular installation, the limit is set to 30
seconds, since the timeout for the HTTP request is 5 seconds we will
instruct the plugin to wait (30 secs - 5 secs) and an additional one
second to spare processing, so in a regular installation the plugin will
try to send as much logs as possible to the API service in less than 25
seconds.