Modify output for the malware results to simplify links
Instead of sending the entire HTML code with the table containing all
the links, iframes and JavaScript files found by the malware scanner,
the plugin will now send the list as a JSON slice, and the HTML code
will be written using the same JavaScript code that executes the Ajax
request. This reduces the size of the object and simplifies the code.