Add option to override the timezone for the datetime
This option defines the timezone that will be used through out the
entire plugin to print the dates and times whenever is necessary. This
option also affects the date and time of the logs visible in the audit
logs panel which is data that comes from a remote server configured to
use Eastern Daylight Time (EDT). WordPress offers an option in the
general settings page to allow you to configure the timezone for the
entire website, however, if you are experiencing problems with the time
in the audit logs, this option will help you fix them.