Add maximum execution time avoidance in the integrity tool
In some environments, the WordPress integrity tool may report a
significant amount of corrupt files, the user will promptly select them
all and execute the "restore" action but due to the way the web server
was configured the operation will fail because the PHP interpreter is
not able to process so many files before the maximum execution time is
reached. This commit fixes this by monitoring the elapsed time and
breaking the loop before the timeout; if this happen the plugin will
display an error message warning the user that the operation reached its
maximum execution time and only a handful of files were processed.