Add option to set WordPress GitHub as the integrity repository
The webmaster can change this URL using an option form the settings page.
This allows them to control which repository will be used to check the
integrity of the installation.

For example, projectnami.org offers an option to use Microsoft SQL Server
instead of MySQL has a different set of files and even with the same
filenames many of them have been modified to support the new database
engine, since the checksums are different than the official ones the
number of false positives will increase. This option allows the webmaster
to point the plugin to a different URL where the new checksums for this
project will be retrieved.

If the custom API is part of GitHub infrastructure, the plugin will try
to build the expected JSON object from the output, if it fails it will
pass the unmodified response to the rest of the code and try to analyze
the integrity of the installation with that information.