Add support to run diff on deleted WordPress files
Right now, the diff utility cannot be used with deleted core files
because the tool needs to compare two existing files, otherwise it
returns an empty result. This commit changes this by creating an empty
temporary file that will act as the modified version of the code, this
way the diff utility will be able to highlight the deleted content.