Modify event report to prefer the custom queue system
When an event is triggered by WordPress the plugin will send the report
to a local queue system which will be dequeued every 24 hours when it
sends the data to the remote API service. The number of logs that will
be send to the API per batch depends on the maximum execution time
allowed by the web server, in a normal configuration, when the maximum
execution time is 30 seconds the plugin will send approximately 15,000
HTTP requests, of course, only if there are +15,000 logs in the queue.