Modify timing for the execution of the Ajax requests
Some web browsers block the rendering of the interface during the
execution of some Ajax requests, this can be prevented by delaying the
execution of the code until the entire interface is loaded and parsed.
With jQuery, for example, this can be accomplished by using the document
ready event as demonstrated in this commit.