Remove unnecessary option queries to the database
During the execution of the custom getOption method the plugin reads the
flat settings file and tries to find the requested option from the JSON
object, if this operation fails the plugin continues with a query to the
database, and if this operation fails the plugin returns a hardcoded
value from the default settings.

This poses a problem when the getOption method is executed multiple
times to return an option that doesn't exists in the flat settings file.
With this commit the plugin will reduce the number of queries to the
database by writing the default value of the non-existing option into
the settings file, this will force the method to return with the first
procedure and skip the interaction with the database.