Fix queries to the database in the last logins page
Multiple parts of the last logins page require the execution of a
functoin that returns the metadata associated to an user account. This
causes a significant number of queries that could be simplified with a
single SQL statemnet to retrieve all the data associated to a list of
user IDs.

This commit implements a patch to collect the unique user identifiers,
execute a single SQL statement against the database, and merge the data
of the existing user accounts into the last logins data set, reducing
the consumption of resources and the load time of this page.