Remove duplicated failed user authentication log
When the plugin detects that the number of failed logins is higher than
the limit set to consider the event as part of a brute force password
guessing attack, it will take all the data collected so far and send it
to the admin users as a report and then reset the content of the log
file. However, the current version of the code is adding the data for
the current event to the flat file twice, once before the report is sent
and again when the log file is reset, this commit fixes this.