Fix audit log parser for incompatible JSON data
Older versions of the code used to store the security events into a
local queue different to the one that is currently implemented. This
queue was used as a fallback for when the API service was down, and so
each log contained the timestamp of the failure which would be used to
simulate the real time of the event on the retry of the HTTP request.

When the current queue system was implemented, this small details was
forgotten and some websites broke because the old logs, which were
encoded using a JSON object, were not supported anymore. This commit
fixes this by ignoring these logs all together.