Add API key in admin notice when it is being deleted
The plugin will print the API key that is being deleted in the dashboard
alert. This allows the user to reinstall the key if the deletion was not
intentional. The user can follow the instructions in the "API key" area
to learn how to insert the deleted key again into the settings file.