Fix modification date for corrupt core files
When the integrity tool prints the list of corrupt files, it includes
the possible modification date for each file. However, for deleted files
this modification date doesn't exists and so the plugin is currently
printing the date and time of the first Unix timestamp second. This
commit fixes this by hiding the modification date for deleted files.