Remove internationalization support for consistency
Because Gettext requires you to use generic text as the key of the
translations, and the majority of text in this project contains long
lines, specially in the templates, I designed a mechanism to allow for
the translation of the entire project into different languages using
unique identifiers as the key for the POT files.

However, after some reports from users in the WordPress forums I found
that this approach was conflicting with some popular plugins which were
using the unique IDs to generate a new POT file into a different folder
in the installation to create new a new translation file. This new file
contains the same value as the key, so the interface was impossible to
use without the real text for the descriptions and buttons.

Also, because the community didn't really contribute to the translation
of the plugin during the time the POT files were available, I decided to
revert the internationalization support and leave the plugin in English
as it was before. This helps us to maintain consistency and reduces the
amount of work required to keep the language files updated.