Modify error message displaying for invalid CSRF validations
When an user submits a form associated to the plugin, an unique random
text is sent along with the rest of the user input, this unique text is
checked using a built-in WordPress function to verify that the data
comes from a valid form rather than a script used to automate form
submissions.

However, this mechanism fails when other add-on popullates the global
POST variable with additional data even when the request is coming from
GET instead of POST. Since educating the users about the causes of this
issue is pointless, and forcing other WordPress developers to follow
good development practices is almost impossible, I have decided to
remove the exit statement and change it with a normal error message
alert. This way the users will still be informed, not only about the
invlaid form submission, but also about the incompatibility with the
other installed add-ons.