Modify option to force scanner to ignore directories
In some cases, the file system scanner will not be able to run because
the website has too many files, and during the limited amount of time
allowed by the server the script will timeout. This option allows the
user to define a list of directories that will be ignored by the
scanner, this could be a cache folder, a media folder, or anything
similar.