Modify internal response to the log file not found error
Some users have reported an intermittent deletion of the API key from
their website after sudden disconnections with the API service. Instead
of deleting the API key thinking that it was invalid we will now log the
error into the local server and hope that the key is valid so we can
send the report back to the API when the connection can be restablished.