Add option to force the firewall cache flush
The plugin allows the admin users to activate the automatic reset of the
firewall cache when a post or page is published. However, this option is
disabled by default to reduce the disruption of the website as the reset
of the cache forces the redownload of all the content and resources
linked to the site that is being visualized. Some users may have a
better configuration and so we will provide the option to override the
default behavior by defining a constant SUCURI_CLEAR_CACHE_ON_PUBLISH in
the configuration file.