Add button to toggle the visibility of the post-types table
In some websites with a lot of plugins, the admin user is having a hard
time to explore the settings page because the list of post-types
available in the alerts panel is too big. When the admin user is not
trying to whitelist or blacklist these post-types it makes no sense to
show them. This commit hides the whole table and offers a button to
toggle the visibility.