Remove option to log passwords during a failed login
This feature, added some time ago after multiple requests from the
WordPress community, has been reviewed again this year by our security
team and we determined that the benefits of knowing the password using
during a failed login attempt are not enough to justify having such a
dangerous logging mechanism.

The scenario used to justify the existence of this feature was that, if
a malicious user was trying to brute force the account of one of the
administrators, one of them could take a look at the logs and determine
if the malicious user was getting close or not to the real password, in
which case would decide to change it.

However, after the security review, we found that knowing the password
used during a failed login attempt is not really useful and on the
contrary opens a hole in the application as a legit admin user could
mistakingly type the semi-correct password in the login form, the plugin
will log the wrong credentials, and if the API key has been leaked, a
malicious user could take this information to access the admin dashboard
by making educated guesses over the mistyped password.