Modify order of the added, modified, removed core files
In the panel that prints the list of added, modified and removed files
to verify the integrity of the WordPress installation, the current code
is not keeping consistency on output, making the testing of this feature
a bit difficult. To increase the coverage of the automated tests, we are
adding a couple of `sort` function calls.