Fix relative file path when ABSPATH is point to root
When the website is installed in the root of the disk, the ABSPATH
constant contains the value `/` which after passing through some
functions that clean the file path turns into an empty string. This
causes the listing in the WordPress integrity panel to print the file
paths without a directory separator.