Remove unused option to reduce number of failed logins
Some time ago some users requested the implementation of a new option to
reduce the number of failed logins associated to an user account that
doesn't exists. This was implemented but later on we discovered that not
many users were taking advantage of this new feature. To reduce the
effort that we have to put into the maintenance of the code, we have
taken the decision to delete this feature for good.