Fix open_basedir restriction was not considered on scans
The execution of functions like `is_dir` or `RecursiveDirectoryIterator`
in an server environment where the `open_basedir` PHP setting has been
set, miserably fails with an unrecoverable warning. This commit adds the
necessary conditions to prevent these warnings from being triggered by
checking if the directory is in the `open_basedir` setting first.