Remove firewall API key deletion on re-authentication
When the user re-generates the API key in their Sucuri Firewall account,
the plugin has no way to know this. So far, the plugin has been reacting
to the re-authentications with the deletion of the previous key without
warning the user about the invalid credentials. Some people complained
about this design and so I have decided to stop the deletion of the key
and add a more detailed error message.