Add option to refresh the sitecheck malware scan cache
The plugin executes the malware scan via SiteCheck every 21,600 seconds
which roughly translates to six hours, using WordPress Task Scheduler.
However, the user may apply changes in the website to react against an
infection and after these changes may want to see an updated version of
the results from the scanner. Waiting several hours is not an option for
many, so here is a simple implementation of an option that will allow
the users to expire the local cache and trigger an API call to SiteCheck
to execute a fresh scan.