Add additional type check in custom array implode
When the user visits “Settings > Scanner > Scheduled Tasks” the plugin
calls a function provided by WordPress called “wp_get_schedules” which
is supposed to return a list with all the active scheduled tasks (aka.
cronjobs). According to the documentation, each item in the list has an
attribute with the arguments that will be passed to the corresponding
functions that will run once the cronjob is executed. This attribute is
supposed to be an array, but in some cases the value uses a different
data type which breaks the implode function.