Modify scheduled tasks panel to load the table via Ajax
There have been cases where a website contains dozens of scheduled tasks
which affect the load time of the settings page because the interpreter
has to wait for all the items in the cronjob list to be processed before
the page is rendered. This commit fixes this problem by loading the data
via Ajax.